requirements.txt
# python==3.10
matplotlib==3.10.3
opencv==4.12.0
pytorch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformers==4.56.1 
monai==1.5.0
accelerate==1.10.1
optuna==4.5.0
scipy==1.15.2
scikit-learn==1.7.2
torchmetrics==1.8.2
scikit-image==0.25.2